Add interface to access score function
This should be used to calculate DQ score/flags for a given DQTree.
The idea is to keep the monitoring tree and the DQ flags separate.
This is required since the DQ flags depend on the Velo state.  So a
particular monitoring tree can have a different set of score/flags for
a different state.

NB: Maybe it is more appropriate to rename DQTree to MoniTree.