Statistical comparison tests with real histograms
Choosing references is not trivial.  If you look at test-statcomp.py,
you will notice tests for cluster size histograms are commented out (and
they are expected to fail).  The problem is, the error on a distribution
from a long good run is very small (~ 1/√n).  This means, if the data
histogram deviates from this "golden standard" a slightest bit, tests
fail (both χ² and KS).

To understand this think of χ² comparison;

  χ² = ∑ᵢ (xᵢ - yᵢ)²/(σxᵢ² + σyᵢ²)

Obviously when the errors are negligible, χ² blows up and the test
fails.  You can see this in action on pages 7 & 8 here:
<https://www.nikhef.nl/~sali/VeloDQ/statistical_comparison_test.pdf>.
In these (normalised) histograms, the y-axis is ~ 10⁶, so the errors are
~ 1/10³; which is very small!

The resolution is to build reference histograms from a wider sample of
several such long, good runs[1]: one entry per run.  Then the reference
will end up having a true (realistic) spread.  We can of course have
different references for different beam conditions/detector
configuration (read: expected state).

Footnotes:

[1] I think I proposed something like this sometime back, can't find a
    reference now.