veloview: first attempt at GRFIO (not working)
Flattening of dictionaries work, but filling the tree doesn't.  I'm
having problems iterating over the keys and filling the appropriate
branches.

My understanding is:

  tree.keyname[timestamp]	# spəcific timestamp
  tree.keyname			# last timestamp

should give me access to the object.  So for writing, I tried this in a
loop:

  getattr(tree, key)[timestamp] = value

But that fails.  I guess the reason is because there are no entries yet
in the branch.  I should use setattr(..) then, but not sure how I can
pass the timestamp key to setattr.