Use a patched version of rq to fix job data bug.
There was an intermittent problem of jobs changing their status to
complete before the job’s output was saved to the Job instance.

This commit uses a fork of mine with a cherry-picked commit from the rq
master branch that fixes this problem applied to v0.4.6.

When v0.4.7 or later arrives, we should check the fix is still present
and update back to tagged rq releases.